Remove warnings

+++ src/core/kernel/qtimerinfo_unix.cpp
@@ -262,7 +262,7 @@ static void calculateCoarseTimerTimeout(QTimerInfo_Unix *t, timespec currentTime
       }
 
    } else {
-      uint min = qMax(0, msec - static_cast<int>(absMaxRounding));
+      uint min = qMax(0, static_cast<int>(msec) - static_cast<int>(absMaxRounding));
       uint max = qMin(1000u, msec + absMaxRounding);
 
       // find the boundary that we want, according to the rules above extra rules:
